refactor: domain 별 구조 변경
