--- src/main/java/com/moodmate/controller/DiaryController.java
@@ -1,28 +1,0 @@
-package com.moodmate.controller;
-
-import com.moodmate.dto.DiarySaveRequestDto;
-import com.moodmate.oauth.CustomOauth2User;
-import com.moodmate.service.DiaryService;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/diaries")
-public class DiaryController {
-
-    private final DiaryService diaryService;
-
-    @PostMapping
-    public ResponseEntity<Long> saveDiary(@RequestBody @Valid DiarySaveRequestDto dto,
-                                          @AuthenticationPrincipal CustomOauth2User userDetails) {
-        Long savedId = diaryService.saveDiary(dto, userDetails.getUser().getId());
-        return ResponseEntity.ok(savedId);
-    }
-}
