+++ src/main/java/com/moodmate/domain/diary/DiaryController.java
@@ -1,0 +1,27 @@
+package com.moodmate.domain.diary;
+
+import com.moodmate.domain.diary.dto.DiarySaveRequestDto;
+import com.moodmate.domain.user.ouath.CustomOauth2User;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/diaries")
+public class DiaryController {
+
+    private final DiaryService diaryService;
+
+    @PostMapping
+    public ResponseEntity<Long> saveDiary(@RequestBody @Valid DiarySaveRequestDto dto,
+                                          @AuthenticationPrincipal CustomOauth2User userDetails) {
+        Long savedId = diaryService.saveDiary(dto, userDetails.getUser().getId());
+        return ResponseEntity.ok(savedId);
+    }
+}
