+++ src/main/java/com/moodmate/domain/diary/DiaryRepository.java
@@ -1,0 +1,10 @@
+package com.moodmate.domain.diary;
+
+import com.moodmate.domain.diary.entity.Diary;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface DiaryRepository extends JpaRepository<Diary, Long> {
+    List<Diary> findByUserId(Long userId);
+}
