+++ src/main/java/com/moodmate/domain/emotion/Emotion.java
@@ -1,0 +1,23 @@
+package com.moodmate.domain.emotion;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Emotion {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true, nullable = false)
+    private String name;
+
+    public Emotion(String name) {
+        this.name = name;
+    }
+}
+
