+++ src/main/java/com/moodmate/domain/emotion/dto/EmotionResponseDto.java
@@ -1,0 +1,11 @@
+package com.moodmate.domain.emotion.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class EmotionResponseDto {
+    private Long id;
+    private String name;
+}
