+++ src/main/java/com/moodmate/domain/user/UserRepository.java
@@ -1,0 +1,13 @@
+package com.moodmate.domain.user;
+
+import com.moodmate.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByLoginId(String loginId);
+    Optional<User> findByEmail(String email);
+    boolean existsByLoginId(String loginId);
+
+}
