+++ src/main/java/com/moodmate/domain/user/entity/Role.java
@@ -1,0 +1,5 @@
+package com.moodmate.domain.user.entity;
+
+public enum Role {
+    USER, ADMIN
+}
