+++ src/main/java/com/moodmate/domain/user/ouath/GoogleUserDetails.java
@@ -1,0 +1,35 @@
+package com.moodmate.domain.user.ouath;
+
+import lombok.AllArgsConstructor;
+
+import java.util.Map;
+
+@AllArgsConstructor
+public class GoogleUserDetails implements OAuth2UserInfo {
+    private Map<String, Object> attributes;
+
+    @Override
+    public String getProvider() {
+        return "google";
+    }
+
+    @Override
+    public String getProviderId() {
+        return (String) attributes.get("sub");
+    }
+
+    @Override
+    public String getEmail() {
+        return (String) attributes.get("email");
+    }
+
+    @Override
+    public String getName() {
+        return (String) attributes.get("name");
+    }
+
+    @Override
+    public String getPicture() {
+        return (String) attributes.get("picture");
+    }
+}
