+++ src/main/java/com/moodmate/domain/user/ouath/OAuth2UserInfo.java
@@ -1,0 +1,10 @@
+package com.moodmate.domain.user.ouath;
+
+public interface OAuth2UserInfo {
+    String getProvider();
+    String getProviderId();
+    String getEmail();
+    String getName();
+
+    String getPicture();
+}
