+++ src/main/java/com/moodmate/domain/user/service/CustomOauth2UserService.java
@@ -1,0 +1,41 @@
+package com.moodmate.domain.user.service;
+
+import com.moodmate.domain.user.entity.User;
+import com.moodmate.domain.user.ouath.CustomOauth2User;
+import com.moodmate.domain.user.ouath.GoogleUserDetails;
+import com.moodmate.domain.user.ouath.OAuth2UserInfo;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
+import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class CustomOauth2UserService extends DefaultOAuth2UserService {
+
+    private final UserService userService;
+
+    @Override
+    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
+        OAuth2User oAuth2User = super.loadUser(userRequest);
+        log.info("getAttributes : {}", oAuth2User.getAttributes());
+
+        String provider = userRequest.getClientRegistration().getRegistrationId();
+        OAuth2UserInfo userInfo = null;
+
+        if (provider.equals("google")) {
+            userInfo = new GoogleUserDetails(oAuth2User.getAttributes());
+        } else if (provider.equals("kakao")) {
+            //userInfo = new KakaoUserDetails(oAuth2User.getAttributes());
+        } else {
+            throw new OAuth2AuthenticationException("지원하지 않는 로그인 제공자입니다.");
+        }
+
+        User user = userService.findOrCreateUser(userInfo, provider);
+        return new CustomOauth2User(user, oAuth2User.getAttributes());
+    }
+}
