--- src/main/java/com/moodmate/dto/EmotionController.java
@@ -1,24 +1,0 @@
-package com.moodmate.dto;
-
-import com.moodmate.service.EmotionService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/emotions")
-public class EmotionController {
-
-    private final EmotionService emotionService;
-
-    @GetMapping
-    public ResponseEntity<List<EmotionResponseDto>> getEmotions() {
-        return ResponseEntity.ok(emotionService.getAllEmotions());
-    }
-}
-
