--- src/main/java/com/moodmate/dto/EmotionResponseDto.java
@@ -1,11 +1,0 @@
-package com.moodmate.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class EmotionResponseDto {
-    private Long id;
-    private String name;
-}
