--- src/main/java/com/moodmate/entity/Emotion.java
@@ -1,23 +1,0 @@
-package com.moodmate.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class Emotion {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true, nullable = false)
-    private String name;
-
-    public Emotion(String name) {
-        this.name = name;
-    }
-}
-
