--- src/main/java/com/moodmate/oauth/OAuth2UserInfo.java
@@ -1,10 +1,0 @@
-package com.moodmate.oauth;
-
-public interface OAuth2UserInfo {
-    String getProvider();
-    String getProviderId();
-    String getEmail();
-    String getName();
-
-    String getPicture();
-}
