--- src/main/java/com/moodmate/oauth/Role.java
@@ -1,5 +1,0 @@
-package com.moodmate.oauth;
-
-public enum Role {
-    USER, ADMIN
-}
