--- src/main/java/com/moodmate/repository/DiaryRepository.java
@@ -1,10 +1,0 @@
-package com.moodmate.repository;
-
-import com.moodmate.entity.Diary;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface DiaryRepository extends JpaRepository<Diary, Long> {
-    List<Diary> findByUserId(Long userId);
-}
