--- src/main/java/com/moodmate/repository/EmotionRepository.java
@@ -1,10 +1,0 @@
-package com.moodmate.repository;
-
-import com.moodmate.entity.Emotion;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface EmotionRepository extends JpaRepository<Emotion, Long> {
-    Optional<Emotion> findByName(String name);
-}
