--- src/main/java/com/moodmate/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.moodmate.repository;
-
-import com.moodmate.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByLoginId(String loginId);
-    Optional<User> findByEmail(String email);
-    boolean existsByLoginId(String loginId);
-
-}
