--- src/main/java/com/moodmate/util/EmotionInitializer.java
@@ -1,26 +1,0 @@
-package com.moodmate.util;
-
-import com.moodmate.entity.Emotion;
-import com.moodmate.repository.EmotionRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-
-@Component
-@RequiredArgsConstructor
-public class EmotionInitializer implements CommandLineRunner {
-
-    private final EmotionRepository emotionRepository;
-
-    @Override
-    public void run(String... args) {
-        List<String> defaultEmotions = List.of("기쁨", "슬픔", "분노", "우울", "뿌듯", "놀람");
-
-        for (String name : defaultEmotions) {
-            emotionRepository.findByName(name)
-                    .orElseGet(() -> emotionRepository.save(new Emotion(name)));
-        }
-    }
-}
